+++ CHANGELOG.md
@@ -65,6 +65,8 @@ For instructions on upgrading to newer versions, visit
 * Calling `Document#as_document` on a frozen document on Rubinius returns the
   attributes instead of nil.
 
+* \#1546 Dirty changes should be returned in a hash with indifferent access.
+
 * \#1542 Eager loading now respects the options (ie skip, limit) provided to
   the criteria when fetch the associations.